+++ cookbooks/vm/attributes/default.rb
@@ -1,4 +1,0 @@
-
-node.default['devbox']['user'] = 'vagrant'
-node.default['devbox']['group'] = 'vagrant'
-node.default['devbox']['password'] = 'vagrant'