+++ cookbooks/vm/recipes/default.rb
@@ -1,6 +1,5 @@
 
 include_recipe 'vm::base'
-include_recipe 'vm::user'
 include_recipe 'vm::bash'
 include_recipe 'vm::git'
 include_recipe 'vm::chefdk'