+++ cookbooks/vm/spec/integration/recipes/bash_spec.rb
@@ -2,12 +2,12 @@ require 'spec_helper'
 
 describe 'vm::bash' do
 
-  it 'sets up the .bash.d/ directory for the devbox user' do
-    expect(file('/home/vagrant/.bash.d')).to be_directory
-    expect(file('/home/vagrant/.bash.d')).to be_owned_by 'vagrant'
+  it 'sets up the .bash.d/ directory for the vm user' do
+    expect(file("#{vm_user_home}/.bash.d")).to be_directory
+    expect(file("#{vm_user_home}/.bash.d")).to be_owned_by vm_user
   end
 
-  it 'sets up shell alias `be` => `bundle exec` for the devbox user' do
-    expect(devbox_user_command('be').stdout).to contain 'bundler: exec needs a command to run'
+  it 'sets up shell alias `be` => `bundle exec` for the vm user' do
+    expect(vm_user_command('be').stdout).to contain 'bundler: exec needs a command to run'
   end
 end