+++ cookbooks/vm/spec/spec_helper.rb
@@ -2,14 +2,24 @@ require 'serverspec'
 
 set :backend, :exec
 
+# the user account under which the VM is set up
+def vm_user
+  'user'
+end
+
+# the configured vm user's home directory
+def vm_user_home
+  "/home/#{vm_user}"
+end
+
 # run the given command in the same environment as if you were logged in to the VM
-def devbox_user_command(cmd)
-  command("sudo -u vagrant bash -i -c '#{cmd}; exit $?'")
+def vm_user_command(cmd)
+  command("sudo -u #{vm_user} bash -i -c '#{cmd}; exit $?'")
 end
 
 # for runnig commands which expect an X environment
-def devbox_user_gui_command(cmd)
-  devbox_user_command "xvfb-run #{cmd}"
+def vm_user_gui_command(cmd)
+  vm_user_command "xvfb-run #{cmd}"
 end
 
 # see https://github.com/sethvargo/chef-sugar/blob/v3.4.0/lib/chef/sugar/docker.rb#L31