+++ CHANGELOG.md
@@ -35,6 +35,10 @@ Improvements:
  * configure PS1 to provide a usable shell prompt for Git (see [PR #45](https://github.com/tknerr/linus-kitchen/pull/45))
  * tweak Atom to scale the tree view's font size along with the editor (see [PR #49](https://github.com/tknerr/linus-kitchen/pull/49))
 
+Breaking changes:
+
+ * the VM is now provisioned under a dedicated user account 'user', and thus the 'vagrant' user account can now be removed before packaging the box (see [PR #44](https://github.com/tknerr/linus-kitchen/pull/44))
+
 ## 0.1 (May 11, 2016)
 
 The initial release of the linus-kitchen developer VM.